Created separate userApps class
