+++ src/app.ts
@@ -1,14 +1,7 @@
 /* eslint-disable camelcase */
 import { Name, PrivateKey, PublicKey } from '@greymass/eosio';
-import { ES256KSigner, createJWT, verifyJWT, JWTVerified } from 'did-jwt';
 import { IDContract } from './services/contracts/IDContract';
-import { KeyManager, KeyManagerLevel } from './services/keymanager';
-import { createStorage, PersistentStorageClean, StorageFactory } from './services/storage';
-import { generateRandomKeyPair, randomString } from './util/crypto';
-import { User } from './user';
-import { createKeyManagerSigner, createSigner } from './services/eosio/transaction';
-import { SdkErrors, throwError } from './services/errors';
-import { createJWK, resolve, toDid } from './util/did-jwk';
+import { createSigner } from './services/eosio/transaction';
 import { getSettings } from './settings';
 import { AccountType, TonomyUsername } from './services/username';
 const idContract = IDContract.Instance;
@@ -47,28 +40,18 @@ namespace AppStatus {
     }
 }
 
-type AppRecord = {
-    account: string;
-    added: Date;
-    status: AppStatus;
-};
+export { AppStatus };
 
-type UserAppStorage = {
-    apps: AppRecord[];
-};
-
-// TODO change to use VC
-type JWTLoginPayload = {
-    randomString: string;
+export interface AppData {
+    accountName: Name;
+    appName: string;
+    username: TonomyUsername;
+    description: string;
+    logoUrl: string;
     origin: string;
-    publicKey: string;
-    callbackPath?: string;
-};
-
-type OnPressLoginOptions = {
-    callbackPath: string;
-    redirect?: boolean;
-};
+    version: number;
+    status: AppStatus;
+}
 
 export type AppCreateOptions = {
     usernamePrefix: string;
@@ -79,24 +62,33 @@ export type AppCreateOptions = {
     publicKey: PublicKey;
 };
 
-export default class App {
-    user: User;
-    keyManager: KeyManager;
-    storage: UserAppStorage & PersistentStorageClean;
+export class App implements AppData {
+    accountName: Name;
+    appName: string;
+    username: TonomyUsername;
+    description: string;
+    logoUrl: string;
+    origin: string;
+    version: number;
+    status: AppStatus;
 
-    constructor(_user: User, _keyManager: KeyManager, storageFactory: StorageFactory) {
-        this.user = _user;
-        this.keyManager = _keyManager;
-        this.storage = createStorage<UserAppStorage>('tonomy.user.app.', storageFactory);
+    constructor(options: AppData) {
+        this.accountName = options.accountName;
+        this.appName = options.appName;
+        this.username = options.username;
+        this.description = options.description;
+        this.logoUrl = options.logoUrl;
+        this.origin = options.origin;
+        this.version = options.version;
+        this.status = options.status;
     }
 
-    static async create(options: AppCreateOptions) {
+    static async create(options: AppCreateOptions): Promise<App> {
         const username = new TonomyUsername(options.usernamePrefix, AccountType.APP, getSettings().accountSuffix);
 
         // TODO remove this
         const privateKey = PrivateKey.from('PVT_K1_2bfGi9rYsXQSXXTvJbDAPhHLQUojjaNLomdm3cEJ1XTzMqUt3V');
 
-        // TODO update storage with app in a pending state
         const res = await idContract.newapp(
             options.appName,
             options.description,
@@ -109,93 +101,16 @@ export default class App {
 
         const newAccountAction = res.processed.action_traces[0].inline_traces[0].act;
 
-        // TODO update status of app to READY or something
-        return {
+        return new App({
             accountName: Name.from(newAccountAction.data.name),
-            username: username,
+            username,
             ...options,
-        };
+            version: newAccountAction.data.version,
+            status: AppStatus.READY,
+        });
     }
 
-    async loginWithApp(account: Name, key: PublicKey, password: string): Promise<void> {
-        const myAccount = await this.user.storage.accountName;
-
-        const appRecord: AppRecord = {
-            account: account.toString(),
-            added: new Date(),
-            status: AppStatus.PENDING,
-        };
-
-        let apps = await this.storage.apps;
-        if (!apps) {
-            apps = [];
-        }
-        apps.push(appRecord);
-        this.storage.apps = apps;
-        await this.storage.apps;
-
-        const signer = createKeyManagerSigner(this.keyManager, KeyManagerLevel.PASSWORD, password);
-        await idContract.loginwithapp(myAccount.toString(), account.toString(), key, signer);
-
-        appRecord.status = AppStatus.READY;
-        this.storage.apps = apps;
-        await this.storage.apps;
-    }
-
-    static async onPressLogin({ redirect = true, callbackPath }: OnPressLoginOptions): Promise<string | void> {
-        const { privateKey, publicKey } = generateRandomKeyPair();
-        const payload: JWTLoginPayload = {
-            randomString: randomString(32),
-            origin: window.location.origin,
-            publicKey: publicKey.toString(),
-            callbackPath,
-        };
-
-        // TODO store the signer key in localStorage
-        const signer = ES256KSigner(privateKey.data.array, true);
-
-        const jwk = await createJWK(publicKey);
-
-        const issuer = toDid(jwk);
-
-        // TODO use expiresIn to make JWT expire after 5 minutes
-        const token = await createJWT(payload, { issuer, signer, alg: 'ES256K-R' });
-
-        if (redirect) {
-            window.location.href = `${getSettings().ssoWebsiteOrigin}/login?jwt=${token}`;
-            return;
-        }
-        return token;
-    }
-
-    static async onRedirectLogin(): Promise<JWTVerified> {
-        const urlParams = new URLSearchParams(window.location.search);
-        const jwt = urlParams.get('jwt');
-        if (!jwt) throwError('No JWT found in URL', SdkErrors.MissingParams);
-
-        const verified = await this.verifyLoginJWT(jwt);
-        const payload = verified.payload as JWTLoginPayload;
-
-        const referrer = new URL(document.referrer);
-        if (payload.origin !== referrer.origin) {
-            throwError(
-                `JWT origin: ${payload.origin} does not match referrer: ${document.referrer}`,
-                SdkErrors.WrongOrigin
-            );
-        }
-        return verified;
-    }
-
-    static async verifyLoginJWT(jwt: string): Promise<JWTVerified> {
-        const resolver: any = {
-            resolve,
-            // TODO add Antelope resolver as well
-        };
-        const res = await verifyJWT(jwt, { resolver });
-
-        if (!res.verified) throwError('JWT failed verification', SdkErrors.JwtNotValid);
-        return res;
-    }
+    // static async from(input: string | Name | TonomyUsername): Promise<App> {
+    //     // TODO
+    // }
 }
-
-export { App, JWTLoginPayload, AppStatus };
