+++ src/index.ts
@@ -4,6 +4,7 @@ import * as Transaction from './services/eosio/transaction';
 export * from './services/keymanager';
 export * from './services/storage';
 export * from './user';
+export * from './userApps';
 export * from './app';
 
 export * from './settings';
@@ -19,4 +20,4 @@ export * from './services/contracts/EosioContract';
 export * from './services/contracts/EosioTokenContract';
 
 export * from './services/errors';
-export * from './app';
+export * from './userApps';
