+++ test/app.test.ts
@@ -1,5 +1,5 @@
 import { PrivateKey, PublicKey } from '@greymass/eosio';
-import App from '../src/app';
+import UserApps from '../src/userApps';
 import { generateRandomKeyPair } from '../src/util/crypto';
 import URL from 'jsdom-url';
 
@@ -11,12 +11,12 @@ describe('logging in', () => {
     });
 
     it('on press button', async () => {
-        const jwt = await App.onPressLogin({ callbackPath: '/login', redirect: false });
+        const jwt = await UserApps.onPressLogin({ callbackPath: '/login', redirect: false });
         expect(jwt).toBeDefined();
     });
 
     it('checks login url', async () => {
-        const jwt = await App.onPressLogin({ callbackPath: '/login', redirect: false });
+        const jwt = await UserApps.onPressLogin({ callbackPath: '/login', redirect: false });
         const url = 'http://localhost/login?jwt=' + jwt;
 
         jest.spyOn(document, 'referrer', 'get').mockReturnValue('http://localhost');
@@ -30,7 +30,7 @@ describe('logging in', () => {
         // @ts-ignore
         global.URL = URL;
 
-        const result = await App.onRedirectLogin();
+        const result = await UserApps.onRedirectLogin();
         expect(result).toBeDefined();
         expect(typeof result.payload.randomString).toBe('string');
         expect(typeof result.payload.publicKey).toBe('string');
